+++ src/index.ts
@@ -2,6 +2,7 @@ import { Type, type Static } from "@sinclair/typebox";
 import { ISO_TIMESTAMP_PATTERN } from "./primitives";
 
 export * from "./contracts";
+export * from "./feeds";
 export * from "./identity";
 export * from "./versioning";
 export {